feat: Rework encoding to start from lower value columns
This allows support for > 2 characters.

The previous technique was attempting to build the braille encoding
from the left to the right. This is problematic when the right most
characters are more than 1 column of '0's. For example, 65536
(= 2 ^ 16) should be '⠁⠀⠀' but once it gets to 2nd from the left
character, it finds that the value it's passing along are 0 and so it
has no means of knowing whether there are 1, 23, or infinity 0s.